Gherkindoc cleanup and some bug fixes for the CLI
